# 06 - Architecture Maps/api-specs/BB-Telemetry.openapi.yaml
openapi: 3.0.3
info:
  title: BB-Telemetry — Unified Micro-USD Consumption Reader
  version: 1.1.0
  description: |
    Reads consumption from both stacks in micro-USD space.
    Sources: (1) genome-system ledger METRIC events, (2) control-brain core.db usage table.
    I3 (fail-closed): if monthly spend exceeds cap_monthly, or billed vs telemetry
    divergence > 20% → FREEZE + [CONFLICT] to human queue + STOP-METABOLIC (death condition).
    Trap #1: "meter or 0" — zero cost with consumed tokens → counted in suspects.
    Trap #2: currency units — costs are USD, caps are AUD, conversion only via pinned FX.
  contact:
    name: Octopus System
  x-invariant:
    - I3: fail-closed → FREEZE on divergence
    - I6: no hardcoded numbers (DIVERGENCE_DEATH from code constant = 0.20, but cap from yaml)
    - I1: append-only ledger
servers: []
paths:
...
  /snapshot:
    get:
      operationId: telemetry.snapshot
      summary: Unified micro-USD snapshot of entire organism + disk write
      description: |
        Combines genome + brain sources. Calculates month and today totals.
        Persists to STATE_DIR/telemetry/ for UI/governor consumption.
      parameters:
        - name: write
          in: query
          required: false
          schema:
            type: boolean
            default: true
      responses:
        "200":
          description: Unified snapshot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TelemetrySnapshot"
      x-python-signature: |
        def snapshot(write: bool = True) -> dict
...
components:
  schemas:
...
    TelemetrySnapshot:
      type: object
      required: [ts, unit, fx_aud_per_usd, sources, genome, brain, per_organ_alltime_musd, month, today, suspect_zero_total]
      properties:
        ts:
          type: string
          format: date-time
        unit:
          type: string
          enum: ["micro-USD (int)"]
        fx_aud_per_usd:
          type: object
          properties:
            rate:
              type: number
            tag:
              type: string
        sources:
          type: object
          properties:
            genome_ledger:
              type: string
              nullable: true
            brain_core_db:
              type: string
              nullable: true
        genome:
          type: object
          properties:
            events:
              type: integer
            cost_musd:
              type: integer
            suspect_zero:
              type: integer
        brain:
          type: object
          properties:
            rows:
              type: integer
            cost_musd:
              type: integer
            suspect_zero:
              type: integer
        per_organ_alltime_musd:
          type: object
          additionalProperties:
            type: integer
        unmapped_musd:
          type: object
          additionalProperties:
            type: integer
        month:
          type: object
          properties:
            key:
              type: string
            musd:
              type: integer
            usd:
              type: number
            aud:
              type: number
        today:
          type: object
          properties:
            musd:
              type: integer
            usd:
              type: number
        suspect_zero_total:
          type: integer
  securitySchemes:
    InternalModule:
      type: apiKey
      in: header
      name: X-Internal-Module
      description: Internal telemetry access only
security:
  - InternalModule: []
...